# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-concurrency flash sale inventory system. Core design: atomic DB UPDATE prevents overselling under concurrent load.\n\n## Concurrency Strategy\nPOST /orders uses a single atomic SQL UPDATE with WHERE guard:\n`UPDATE campaign SET remaining_qty = remaining_qty - $qty WHERE ... AND remaining_qty \u003e= $qty`\nRowsAffected == 0 signals sold-out. No SELECT FOR UPDATE needed.",
        "title": "Flash Sale API",
        "contact": {
            "name": "changken"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/campaigns": {
            "get": {
                "description": "Returns all flash sale campaigns with remaining inventory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "List all campaigns",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CampaignWithProduct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/campaigns/{id}": {
            "get": {
                "description": "Returns a single campaign with product info and remaining inventory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Get campaign by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CampaignWithProduct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "description": "Atomically deducts inventory and creates an order. Core concurrency endpoint — protected by DB atomic UPDATE with WHERE guard.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Place a flash sale order",
                "parameters": [
                    {
                        "description": "Order request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.createOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Out of stock or campaign not active",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Returns a single active product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.createOrderRequest": {
            "type": "object",
            "required": [
                "campaign_id",
                "qty",
                "user_id"
            ],
            "properties": {
                "campaign_id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000002"
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "unique-uuid-from-client"
                },
                "qty": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "user_id": {
                    "type": "string",
                    "example": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
                }
            }
        },
        "handler.createOrderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-03-29T10:00:00Z"
                },
                "order_id": {
                    "type": "string",
                    "example": "d6aa7191-1ffc-4be3-a224-bf8bf6be5ebb"
                },
                "qty": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "subtotal": {
                    "type": "number",
                    "example": 29900
                },
                "unit_price": {
                    "type": "number",
                    "example": 29900
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string",
                    "example": "OUT_OF_STOCK"
                },
                "message": {
                    "type": "string",
                    "example": "Sorry, the campaign items are sold out."
                }
            }
        },
        "model.CampaignStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "CampaignDraft",
                "CampaignActive",
                "CampaignEnded",
                "CampaignCancelled"
            ]
        },
        "model.CampaignWithProduct": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orig_price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "remaining_qty": {
                    "type": "integer"
                },
                "sale_price": {
                    "type": "number"
                },
                "start_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.CampaignStatus"
                },
                "total_qty": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/model.ProductStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "model.ProductStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "ProductActive",
                "ProductInactive"
            ]
        }
    }
}